Changes to 'addr' element in riscv_disasm and riscv_asm
- Add debug version string in riscv-compress-elf
- Remove addr from the riscv_disasm struct and reduce it to 32 bytes
- Add addr to the riscv_asm struct in riscv-compress-elf
- Documented the intention of 'label_cont' in this commit message
  and the rationale behind moving 'addr' to riscv_asm.

The addr element was in riscv_disasm to allow the address decode
functions to be shared with riscv_asm however the disassembler
only required simple address decoding and did not need to tag
the address with relocation types or store the address, as the
disassembler only holds the address transiently to print it,
hence the move of addr from riscv_disasm to riscv_asm.

The riscv_asm struct in riscv-compress-elf will eventually be
promoted to a riscv-asm module when it has enough features.

The continuation label 'label_cont' is presently not being used
however a future version may support an ABI that allows passing
continuation address besides the realtive address following a
jump and link instruction (pc + 4). This allows rewriting jump
and link to {auipc,addi,ld} and a jump or jalr to rewrite either
or both of the return address or target procedure address,
potentially via table indirection to allow far calls. This will
allow code to be reordered around procedure call boundaries and
provide a mechanism to insert stub code that returns to an
interpreter while retaining the original control flow in
'label_cont'. This becomes relevant when blocks are relocated
non-locally as continuations.