Add parse-meta generated file diff test to prevent spurious rebuilds.
Note: the parse_meta make function will generate the source or headers
to temporary files and only update the target source or headers if the
output differs. This is to prevent spurious rebuilds when the generated
source is the same, however due to this check, the prerequisites may be
newer than the output files and these targets will get called each time
make is run as the timestamps are not updated unless the output differs.