Use integral_t to differentiate addresses and integers in config.
Integrals are a subset of scalars. It is not correct to define
the integral type as scalar_t so it has been renamed to integral_t.

addr_t will later be specialised to include the context i.e.
user_addr_t, physical_addr_t, virtual_addr_t however at this point
we want to differentiate addresses and integrals.

Likewise we don't want to use int_t for integral_t. integral_t
is currently a placeholder for the largest integral type. We could
potentially use sxlen_t or uxlen_t (although this can vary between
the host and the emulated guest). Something to think about...

An address is essentially a bit vector with segments specifying
dimensions for routing on an address bus. e.g. a physical address
may have bit components that specify a DIMM, a row and a column,
however for the purposes of configuration we store addresses in
an integral type.