Change fmt to ignore width modifiers and error on long and wide types
- Remove dependency on wcrtomb which is a locale support routine.

The formatter is intended to evolve towards a similar interface to Go.
The intent is to keep the code relatively compact so that it could
be used in a firmware image that requires string formatting.

Support for UTF-8 needs to be handled a layer above, and is only
supported in the formatter by appending valid UTF-8 strings.
Rendering code could still use the output of the formatter for
rendering UTF-8 however the formatter is encoding agnostic.

- long double is not supported
- locales and grouping are not supported
- position arguments of the form *nn$ are not supported
- 'C' format for wchar_t characters are not supported
- 'S' format for wchar_t strings are not supported
- 'O', D', 'U', 'X' or 'L' to represent long types are not supported
- 'hjlqtz' width modifiers are ignored as type is captured
- 'n' format to write back length to a variable is not supported