Added mmap interposition and removed dlmalloc
The executables are linked at a high address to make room for
guest address space. It is possible to link at 0x7ffe00000000
on Linux ELF systems however there appears to be an issue with
older binutils so the link address is 0x7e000000 on Linux.

These are the current linker flags:

 * MACOS_LDFLAGS = -Wl,-pagezero_size,0x1000 -Wl,-no_pie \
                   -image_base 0x7ffe00000000
 * LINUX_LDFLAGS = -Wl,-Ttext-segment=0x7e000000

The dlmalloc custom allocator has been removed in favor of mmap
interceptiond. libSystem on macOS normally allocates the bottom
4GiB of address space as a large zero page but when it is set to
4K libSystem heap allocations collide with the low addresses which
need to be reserved for guest address space. mmap-macos.dylib
has mitigations that prevent the libSystem memory allocator from
colliding with the low 4GiB. This is achieved by overriding
mmap, vm_map, vm_allocate, mach_vm_map and mach_vm_allocate
such that the default address hint causes an allocation in high
memory. libsystem_malloc.dylib uses the address hints to pass
entropy and uses the VM_FLAGS_ANYWHERE flag in mach_vm_allocate
and mach_vm_map. mmap-linux.so interposes mmap in a similar way
for linux.